requirements.txt
Django~=4.0.0
djangorestframework~=3.13.1
requests~=2.26.0
